Refactor: move Project API into /project package for better modularity
