+++ src/main/java/com/david/examportfolio/exam_portfolio_backend/project/exception/ApiErrorResponse.java
@@ -1,4 +1,4 @@
-package com.david.examportfolio.exam_portfolio_backend.exception;
+package com.david.examportfolio.exam_portfolio_backend.project.exception;
 
 import java.util.List;
