+++ src/main/java/com/david/examportfolio/exam_portfolio_backend/project/exception/GlobalExceptionHandler.java
@@ -1,6 +1,5 @@
-package com.david.examportfolio.exam_portfolio_backend.exception;
+package com.david.examportfolio.exam_portfolio_backend.project.exception;
 
-import org.apache.coyote.Response;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
