+++ src/main/java/com/david/examportfolio/exam_portfolio_backend/project/exception/ProjectNotFoundException.java
@@ -1,4 +1,4 @@
-package com.david.examportfolio.exam_portfolio_backend.exception;
+package com.david.examportfolio.exam_portfolio_backend.project.exception;
 
 public class ProjectNotFoundException extends RuntimeException {
