+++ src/main/java/com/david/examportfolio/exam_portfolio_backend/project/exception/ValidationError.java
@@ -1,4 +1,4 @@
-package com.david.examportfolio.exam_portfolio_backend.exception;
+package com.david.examportfolio.exam_portfolio_backend.project.exception;
 
 public record ValidationError(
         String field,
