+++ src/main/java/com/david/examportfolio/exam_portfolio_backend/project/service/ProjectService.java
@@ -1,12 +1,12 @@
-package com.david.examportfolio.exam_portfolio_backend.service;
+package com.david.examportfolio.exam_portfolio_backend.project.service;
 
-import com.david.examportfolio.exam_portfolio_backend.dto.admin.AdminProjectDTO;
-import com.david.examportfolio.exam_portfolio_backend.dto.admin.RequestProjectDTO;
-import com.david.examportfolio.exam_portfolio_backend.dto.user.ResponseProjectDTO;
-import com.david.examportfolio.exam_portfolio_backend.exception.ProjectNotFoundException;
-import com.david.examportfolio.exam_portfolio_backend.mapper.ProjectMapper;
-import com.david.examportfolio.exam_portfolio_backend.model.Project;
-import com.david.examportfolio.exam_portfolio_backend.repository.ProjectRepository;
+import com.david.examportfolio.exam_portfolio_backend.project.dto.admin.AdminProjectDTO;
+import com.david.examportfolio.exam_portfolio_backend.project.dto.admin.RequestProjectDTO;
+import com.david.examportfolio.exam_portfolio_backend.project.dto.user.ResponseProjectDTO;
+import com.david.examportfolio.exam_portfolio_backend.project.exception.ProjectNotFoundException;
+import com.david.examportfolio.exam_portfolio_backend.project.mapper.ProjectMapper;
+import com.david.examportfolio.exam_portfolio_backend.project.model.Project;
+import com.david.examportfolio.exam_portfolio_backend.project.repository.ProjectRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
